# Pick off some (very) low handing fruit in compiler perf area