--- a/edb/ir/scopetree.py
+++ b/edb/ir/scopetree.py
@@ -194,17 +194,17 @@ class ScopeTreeNode:
     @property
     def path_children(self) -> Iterator[ScopeTreeNodeWithPathId]:
         """An iterator of node's children that have path ids."""
-        return cast(
-            Iterator[ScopeTreeNodeWithPathId],
-            filter(lambda p: p.path_id is not None, self.children),
+        return filter(
+            lambda p: p.path_id is not None,
+            self.children,  # type: ignore
         )
 
     @property
     def path_descendants(self) -> Iterator[ScopeTreeNodeWithPathId]:
         """An iterator of node's descendants that have path ids."""
-        return cast(
-            Iterator[ScopeTreeNodeWithPathId],
-            filter(lambda p: p.path_id is not None, self.descendants),
+        return filter(
+            lambda p: p.path_id is not None,
+            self.descendants,  # type: ignore
         )
 
     def get_all_paths(self) -> Set[pathid.PathId]: