A few simple micro-optimisations and caches result in ~20% perf boost.
The baseline is the names-as-tuples branch, which regressed a bit compared
to master (which spurred me to look into clawing some of that back).

    compile_edgeql_script_simple: Mean +- std dev: 1.48 ms += 0.02 ms (6.18% faster)
    compile_edgeql_script_webapp_configure_system: Mean +- std dev: 19.4 ms += 0.2 ms (18.40% faster)
    compile_edgeql_script_webapp_get_movie: Mean +- std dev: 117 ms += 2 ms (19.55% faster)
    compile_edgeql_script_webapp_get_person: Mean +- std dev: 135 ms += 3 ms (22.30% faster)
    compile_edgeql_script_webapp_get_user: Mean +- std dev: 72.4 ms += 0.9 ms (23.68% faster)
    compile_edgeql_script_webapp_insert_movie: Mean +- std dev: 71.3 ms += 1.1 ms (16.42% faster)
    compile_edgeql_script_webapp_insert_person: Mean +- std dev: 20.3 ms += 0.2 ms (16.76% faster)
    compile_edgeql_script_webapp_insert_review: Mean +- std dev: 36.0 ms += 0.6 ms (14.46% faster)